+++ src/pages/Quiz.tsx
@@ -1,13 +1,18 @@
-import React, { useEffect, useRef, useState } from 'react';
-
-import { getQuiz } from '../api/quiz';
-import { styled } from 'styled-components';
-import quizBG from '../images/quiz/quiz_background2.png';
+import React, { useState, useEffect } from 'react';
 import NextButton from '../componants/quiz/NextButton';
 import { useQuery } from '@tanstack/react-query';
-import Button from '../componants/Button';
-import { useNavigate } from 'react-router-dom';
-import { result } from '../resultData/Result';
+import { getQuiz } from '../api/quiz';
+import QuizMainContents from '../componants/quiz/QuizMainContents';
+import {
+  StHillImg,
+  StHillImgWrap,
+  StNextButtonWrap,
+  StQuizMain,
+  StSection,
+  StStarImg,
+  StTitle,
+} from '../style/quiz/MainStyle';
+import QuizResult from '../componants/quiz/QuizResult';
 
 interface IQuizs {
   id: string;
@@ -20,27 +25,18 @@ const hillsImg: string[] = ['Hill1.svg', 'Hill2.svg', 'Hill3.svg', 'Hill4.svg'];
 const starImg: string[] = ['Aquarius.svg', 'Aries.svg', 'Cancer.svg', 'Capricorn.svg'];
 
 const Quiz = () => {
-  const { isLoading, isError, data } = useQuery<IQuizs[]>(['quiz'], getQuiz);
   const [number, setNumber] = useState<number>(-1);
+  const { isLoading, isError, data } = useQuery<IQuizs[]>(['quiz'], getQuiz);
   const [quizs, setQuizs] = useState<IQuizs[]>([]);
   const [quizImg, setQuizImg] = useState<string>('');
   const [quizData, setQuizData] = useState<string[]>([]);
   const [userAnswer, setUserAnswer] = useState<string[]>([]);
   const [answers, setAnswers] = useState<string[]>([]);
   const [score, setScore] = useState<number>(0);
-  const navigate = useNavigate();
 
-  // const clickNextBtnHandler = () => {
-  //   setNumber(number + 1);
-  //   if (quizs[number]?.answer === userAnswer) {
-  //     setScore(score + 1);
-  //   }
-  //   setUserAnswer('');
-  // };
   const clickNextBtnHandler = () => {
     if (number >= 0) {
       answers[number] = quizs[number]?.answer;
-      console.log(answers);
       setAnswers(answers);
     }
 
@@ -48,38 +44,13 @@ const Quiz = () => {
     if (number === 4) {
       let userScore = 0;
       answers.forEach((answer, i) => {
-        console.log('q', answer);
-        console.log('user', userAnswer[i]);
-        console.log(answer === userAnswer[i]);
         if (answer === userAnswer[i]) {
           userScore = userScore + 1;
         }
         setScore(userScore);
       });
     }
-    // if (number > 0) {
-    // }
-    // if (quizs[number]?.answer === userAnswer) {
-    //   setScore(score + 1);
-    // }
-    // setUserAnswer('');
   };
-
-  const clickPrevBtnHandler = () => {
-    setNumber(number - 1);
-  };
-
-  // const inputOnChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-  //   setUserAnswer(e.target.value);
-  //   // console.log(quizs[number]?.answer);
-  //   // console.log(userAnswer);
-  // };
-  const inputOnChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-    let userAnswerArr = [...userAnswer];
-    userAnswerArr[number] = e.target.value;
-    setUserAnswer(userAnswerArr);
-  };
-
   useEffect(() => {
     if (isLoading === true) {
       return;
@@ -124,6 +95,10 @@ const Quiz = () => {
             <p>" 나는 별자리를 얼마나 알고 있을까? "</p>
             <p>총 5문제 - 소요시간 약 5분</p>
           </StTitle>
+          <StNextButtonWrap>
+            <p>Start!!</p>
+            <NextButton type="submit" onClick={clickNextBtnHandler} />
+          </StNextButtonWrap>
           <div>
             {starImg.map((img, i) => {
               return <StStarImg key={img + i} src={`../images/quiz/${img}`} />;
@@ -134,323 +109,29 @@ const Quiz = () => {
               return <StHillImg key={img + i} src={`../images/quiz/${img}`} />;
             })}
           </StHillImgWrap>
-          <StNextButtonWrap>
-            <p>Start!!</p>
-            <NextButton type="submit" onClick={clickNextBtnHandler} />
-          </StNextButtonWrap>
         </StQuizMain>
       ) : number < 5 ? (
-        <StQuizContents>
-          <p>아래 별자리를 보고 알맞은 별자리의 이름을 클릭하세요</p>
-          <p>{`문제 ${number + 1}`}/5</p>
-          <StQuizWrap>
-            <div>
-              <img src={quizImg} alt="별자리 이미지" />
-              {/* {isImageLoaded ? <p>Image loaded successfully!</p> : <p>Image is still loading...</p>} */}
-            </div>
-            <StForm
-              onSubmit={e => {
-                e.preventDefault();
-              }}>
-              {quizData?.map((item, i) => {
-                return (
-                  <div key={item + i}>
-                    <input type="radio" id={item} name="checkAnswer" value={item} onChange={inputOnChangeHandler} />
-                    <label htmlFor={item}>
-                      {i + 1 + '. '}
-                      {item}
-                    </label>
-                  </div>
-                );
-              })}
-              <NextButton className="prevBtn" type="submit" onClick={clickPrevBtnHandler} />
-              {userAnswer[number] && (
-                <StNextButtonWrap>
-                  <p>Next</p>
-                  <NextButton type="submit" onClick={clickNextBtnHandler} />
-                </StNextButtonWrap>
-              )}
-            </StForm>
-          </StQuizWrap>
-        </StQuizContents>
+        <>
+          {userAnswer[number] && (
+            <StNextButtonWrap>
+              <p>Next</p>
+              <NextButton type="submit" onClick={clickNextBtnHandler} />
+            </StNextButtonWrap>
+          )}
+          <QuizMainContents
+            number={number}
+            setNumber={setNumber}
+            userAnswer={userAnswer}
+            setUserAnswer={setUserAnswer}
+            quizData={quizData}
+            quizImg={quizImg}
+          />
+        </>
       ) : (
-        <StQuizResult>
-          <h2>⭐퀴즈 결과⭐</h2>
-          <p>총 5문제 중 {score}문제 맞추셨습니다!</p>
-          <p>{result[score].title}</p>
-          <pre>{result[score].contents}</pre>
-          <img src={`../images/result/result${score + 1}.jpg`} />
-          <StResultButtonWrap>
-            <Button size="medium" onClick={() => window.location.reload()}>
-              다시하기
-            </Button>
-            <Button size="medium" onClick={() => navigate('/')}>
-              메인페이지 이동
-            </Button>
-          </StResultButtonWrap>
-        </StQuizResult>
+        <QuizResult score={score} />
       )}
     </StSection>
   );
 };
 
 export default Quiz;
-const StSection = styled.section`
-  background: url(${quizBG});
-  color: var(--color_white);
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  position: relative;
-`;
-const StQuizMain = styled.div`
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  position: relative;
-`;
-const StTitle = styled.div`
-  position: absolute;
-  text-align: center;
-  animation: slideinTop 2s forwards;
-
-  @keyframes slideinTop {
-    0% {
-      opacity: 0%;
-      top: -60px;
-    }
-    100% {
-      opacity: 100%;
-      top: 250px;
-    }
-  }
-  & h2 {
-    font-size: 36px;
-    font-weight: 800;
-    margin-bottom: 60px;
-  }
-  & p {
-    font-size: 24px;
-    margin: 10px 0;
-    text-align: center;
-    &:nth-child(3) {
-      font-size: 18px;
-      margin-top: 20px;
-    }
-  }
-`;
-const StStarImg = styled.img`
-  opacity: 0;
-  position: absolute;
-  animation-name: fadeInOut;
-  animation-iteration-count: infinite;
-  animation-direction: alternate;
-  animation-duration: 2s;
-  &:nth-child(1) {
-    left: 160px;
-    top: 180px;
-    width: 200px;
-    animation-delay: 2s;
-  }
-  &:nth-child(2) {
-    left: 500px;
-    top: 100px;
-    width: 170px;
-    animation-delay: 4s;
-  }
-  &:nth-child(3) {
-    right: 570px;
-    top: 70px;
-    width: 170px;
-    animation-delay: 3s;
-  }
-  &:nth-child(4) {
-    right: 270px;
-    top: 180px;
-    width: 170px;
-    animation-delay: 5s;
-  }
-  @keyframes fadeInOut {
-    0% {
-      opacity: 0%;
-    }
-    100% {
-      opacity: 100%;
-    }
-  }
-`;
-const StHillImgWrap = styled.div`
-  width: 100%;
-  overflow: hidden;
-  position: relative;
-  height: 100vh;
-`;
-const StHillImg = styled.img`
-  width: 60%;
-  position: absolute;
-  bottom: 0px;
-
-  &:nth-child(odd) {
-    animation-name: slideinLeft;
-    animation-duration: 2s;
-  }
-  &:nth-child(even) {
-    right: 0px;
-    animation-name: slideinRight;
-    animation-duration: 2s;
-  }
-  &:nth-child(1) {
-    width: 80%;
-  }
-  @keyframes slideinLeft {
-    0% {
-      left: -800px;
-    }
-    100% {
-      left: 0px;
-    }
-  }
-  @keyframes slideinRight {
-    0% {
-      right: -800px;
-    }
-    100% {
-      right: 0px;
-    }
-  }
-`;
-const StNextButtonWrap = styled.div`
-  text-align: center;
-  position: absolute;
-  top: calc(50% - 34px);
-  transform: translate(-50%, 0px);
-  animation: moveBtn 0.8s forwards;
-  animation-iteration-count: infinite;
-  animation-direction: alternate;
-
-  @keyframes moveBtn {
-    0% {
-      right: 250px;
-    }
-    100% {
-      right: 260px;
-    }
-  }
-  &:hover {
-    animation-play-state: paused;
-  }
-  & p {
-    margin-bottom: 10px;
-  }
-`;
-
-const StQuizContents = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  & p {
-    font-size: 24px;
-    &:nth-child(2) {
-      margin-top: 20px;
-      font-size: 18px;
-    }
-  }
-`;
-
-const StQuizWrap = styled.div`
-  text-align: center;
-
-  & img {
-    width: 300px;
-    height: 300px;
-    margin: 80px;
-  }
-`;
-
-const StForm = styled.form`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  font-size: 24px;
-  & div {
-    & input {
-      opacity: 0;
-    }
-    & input[type='radio']:checked + label:after {
-      transform: scaleX(1);
-      /* border-bottom: 3px solid var(--color_purple);
-      padding-bottom: 5px; */
-    }
-
-    & label {
-      cursor: pointer;
-      margin: 0px 30px;
-      display: inline-block;
-
-      &:after {
-        display: block;
-        content: '';
-        margin-top: 10px;
-        border-bottom: 5px solid var(--color_purple);
-        transform: scaleX(0);
-        transition: transform 250ms ease-in-out;
-        transform-origin: 0% 50%;
-        opacity: 70%;
-      }
-      &:hover:after {
-        transform: scaleX(1);
-        /* width: 100px; */
-      }
-    }
-  }
-  & .prevBtn {
-    text-align: center;
-    position: absolute;
-    top: 50%;
-    left: 250px;
-    transform: scaleX(-1) translate(-50%, 0px);
-  }
-`;
-const StQuizResult = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  & h2 {
-    font-size: 36px;
-    font-weight: 800;
-    margin-bottom: 30px;
-  }
-  & p {
-    &:nth-child(3) {
-      margin-top: 60px;
-      margin-bottom: 30px;
-      font-size: 24px;
-      font-family: 'RixTodaytoonB';
-    }
-    &:nth-child(4) {
-      text-align: center;
-    }
-  }
-  & pre {
-    text-align: center;
-    line-height: 1.3;
-  }
-  & img {
-    width: 400px;
-    margin: 60px 0;
-  }
-`;
-const StResultButtonWrap = styled.div`
-  display: flex;
-  flex-direction: row;
-  & button:nth-child(2) {
-    margin-left: 30px;
-  }
-`;
